pyproject.toml
[project]
name = "qdrant_rag"
version = "0.1.0"
description = "Multi-tenant Qdrant vector storage service for RAG."
requires-python = ">=3.13"
dependencies = [
    "celery>=5.4",
    "django>=5.2,<6.0",
    "django-environ>=0.13.0",
    "djangorestframework>=3.16",
    "FlagEmbedding>=1.3",
    "gunicorn>=25.3.0",
    "langchain-text-splitters>=0.3",
    "prometheus-client>=0.20",
    "psycopg[binary,pool]>=3.3.3",
    "qdrant-client>=1.17.1",
    "redis>=5.0",
    "structlog>=25.5.0",
    "torch>=2.4",
    "whitenoise>=6.7",
]

[dependency-groups]
dev = [
    "django-stubs[compatible-mypy]>=6.0.3",
    "grpcio-reflection>=1.60",
    "httpx>=0.28.1",
    "mypy>=1.20.2",
    "pytest>=9.0.3",
    "pytest-cov>=7.1.0",
    "pytest-django>=4.12.0",
    "ruff>=0.15.12",
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cpu" }

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.test_settings"
python_files = "test_*.py"
testpaths = ["tests"]
addopts = "-ra -q --strict-markers"
markers = [
    "embedder: tests that require the BGE-M3 model to be loaded (slow, ~30s+ first run)",
]

[tool.ruff]
line-length = 100
target-version = "py313"
extend-exclude = ["apps/grpc_service/generated"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "A", "C4", "SIM"]
ignore = ["E501"]
